fetchSchoolsとfetchGradesをフック化
